+++ NEWS.md
@@ -1,8 +1,13 @@
 # chart News
 
+## chart version 1.1.0
+
+- First implementation of base and lattice versions of `chart()`s.
+
+
 ## chart version 1.0.1
 
-- A bug in `chart()` prodced an errow when labels are not defined and
+- A bug in `chart()` produced an error when labels are not defined and
   `auto.labs = TRUE`.